# k8s-deploy/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a KubeFATE.",
        "title": "KubeFATE service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/FederatedAI/KubeFATE"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1"
    },
    "host": "kubefate.net",
    "basePath": "/v1",
    "paths": {
...
        "/cluster": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "OAuth2Application": [
                            "write",
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Updates a cluster in the store with form data",
                "parameters": [
                    {
                        "description": "Cluster Args",
                        "name": "ClusterArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/job.ClusterArgs"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modules.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "OAuth2Application": [
                            "write",
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Create a new cluster",
                "parameters": [
                    {
                        "description": "Cluster Args",
                        "name": "ClusterArgs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/job.ClusterArgs"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modules.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            }
        },
        "/cluster/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "OAuth2Application": [
                            "write",
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "List all available clusters",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "get All cluster",
                        "name": "all",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/modules.Cluster"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            }
        },
        "/cluster/{clusterId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "OAuth2Application": [
                            "write",
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "create Cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster Id",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modules.Cluster"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "OAuth2Application": [
                            "write",
                            "admin"
                        ]
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cluster"
                ],
                "summary": "Delete a cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "cluster Id",
                        "name": "clusterId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.JSONResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/modules.Job"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized operation",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/api.JSONERRORResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.JSONERRORResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "api.JSONResult": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "job.ClusterArgs": {
            "type": "object",
            "properties": {
                "chart_name": {
                    "type": "string"
                },
                "chart_version": {
                    "type": "string"
                },
                "cover": {
                    "type": "boolean"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "name": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "modules.Cluster": {
            "type": "object",
            "properties": {
                "Info": {
                    "description": "Info is the corresponding information of cluster in k8s",
                    "$ref": "#/definitions/modules.MapStringInterface"
                },
                "Spec": {
                    "description": "Spec corresponding to values(of Cluster) is decoded into interface by yaml",
                    "$ref": "#/definitions/modules.MapStringInterface"
                },
                "chart_name": {
                    "type": "string"
                },
                "chart_values": {
                    "description": "Values through the values in the chart file- template.yaml The file generates the corresponding helm values file",
                    "$ref": "#/definitions/modules.MapStringInterface"
                },
                "chart_version": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "helm_revision": {
                    "description": "Cluster revision of helm",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "namespaces": {
                    "type": "string"
                },
                "revision": {
                    "description": "Cluster revision",
                    "type": "integer"
                },
                "status": {
                    "description": "The status of the cluster, including: \"Creating\",\"Deleting\",\"Updating\",\"Running\",\"Unavailable\",\"Deleted\"",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "description": "UUID is created by github.com/satori/go A combination of 36 bit strings generated by. UUID",
                    "type": "string"
                },
                "values": {
                    "description": "Values field storage cluster.yaml File content of",
                    "type": "string"
                }
            }
        },
...
        "modules.Job": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creator": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "end_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "result": {
                    "type": "string"
                },
                "start_time": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "sub_jobs": {
                    "$ref": "#/definitions/modules.SubJobs"
                },
                "time_limit": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "modules.MapStringInterface": {
            "type": "object",
            "additionalProperties": true
        },
...
        "modules.SubJob": {
            "type": "object",
            "properties": {
                "endTime": {
                    "type": "string"
                },
                "moduleName": {
                    "type": "string"
                },
                "modulesPodStatus": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "modules.SubJobs": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/modules.SubJob"
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}